move get withdrawal balance to its own file
